# src/swagger/openapi.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "ToDoList API",
      "version": "v1"
    },
    "servers": [
      {
        "url": "http://localhost:5002"
      }
    ],
    "tags": [
      {
        "name": "API functions",
        "description": "API functions of our application"
      }
    ],
    "paths": {
      "/to-do-list": {
        "get": {
          "summary": "returns all to",
          "tags": [
            "API functions"
          ],
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "required": true,
              "description": "simple parameter",
              "schema": {
                "type": "number",
                "example": "10"
              }
            },
            {
                "name": "page",
                "in": "query",
                "required": true,
                "description": "simple parameter",
                "schema": {
                  "type": "number",
                  "example": "0"
                }
              },
              {
                "name": "isCompleted",
                "in": "query",
                "required": true,
                "description": "simple parameter",
                "schema": {
                  "type": "boolean",
                  "example": "true"
                }
              }
          ],
          "description": "parameter id just for test",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "document": {
                  "schema": {
                    "type": "string",
                    "example": "some text"
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "returns simple answer from post",
          "tags": [
            "API functions"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                },
                "example": {
                  "country": "Russia",
                  "city": "Spb"
                }
              }
            }
          },
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "simple parameter",
              "schema": {
                "type": "string",
                "example": "1"
              }
            }
          ],
          "description": "parameter id just for test",
          "responses": {
            "204": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Info": {
          "type": "object",
          "properties": {
            "country": {
              "type": "string"
            },
            "city": {
              "type": "string"
            }
          }
        }
      }
    }
  }